+++ CHANGELOG.md
@@ -5,7 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased] 
+## Unreleased
+### Added
+- Method TSDataset.info ([#409](https://github.com/tinkoff-ai/etna/pull/409))
+
+## Changed
+- Change method TSDataset.describe ([#409](https://github.com/tinkoff-ai/etna/pull/409))
 
 ### Fixed
 - Add relevance_params in GaleShapleyFeatureSelectionTransform ([#410](https://github.com/tinkoff-ai/etna/pull/410))